+++ examples/instance_template/simple/main.tf
@@ -23,6 +23,7 @@ provider "google" {
 
 module "instance_template" {
   source          = "../../../modules/instance_template"
+  project_id      = var.project_id
   subnetwork      = var.subnetwork
   service_account = var.service_account
   name_prefix     = "simple"