+++ examples/mig/full/variables.tf
@@ -39,7 +39,6 @@ variable "subnetwork" {
 
 variable "project_id" {
   description = "The GCP project to use for integration tests"
-  default     = ""
 }
 
 variable "credentials_path" {