+++ examples/mig/simple/main.tf
@@ -16,26 +16,27 @@
 
 provider "google" {
   credentials = file(var.credentials_path)
-  project     = var.project_id
   region      = var.region
   version     = "~> 2.7.0"
 }
 
 provider "google-beta" {
   credentials = file(var.credentials_path)
-  project     = var.project_id
   region      = var.region
   version     = "~> 2.7.0"
 }
 
 module "instance_template" {
-  source          = "../../../modules/instance_template"
-  subnetwork      = var.subnetwork
-  service_account = var.service_account
+  source             = "../../../modules/instance_template"
+  project_id         = var.project_id
+  subnetwork         = var.subnetwork
+  service_account    = var.service_account
+  subnetwork_project = var.project_id
 }
 
 module "mig" {
   source            = "../../../modules/mig"
+  project_id        = var.project_id
   region            = var.region
   target_size       = var.target_size
   hostname          = "mig-simple"