+++ modules/preemptible_and_regular_instance_templates/main.tf
@@ -21,6 +21,7 @@
 module "preemptible" {
   source               = "../../modules/instance_template"
   name_prefix          = "${var.name_prefix}-regular"
+  project_id           = var.project_id
   machine_type         = var.machine_type
   labels               = var.labels
   metadata             = var.metadata
@@ -44,6 +45,7 @@ module "preemptible" {
 module "regular" {
   source               = "../../modules/instance_template"
   name_prefix          = "${var.name_prefix}-preemptible"
+  project_id           = var.project_id
   machine_type         = var.machine_type
   labels               = var.labels
   metadata             = var.metadata